feat: fix constants and types in ScalarN, KeyCodec, partial Glv fix (WIP — still broken)
Progress on the LongArray(4) migration:
- ScalarN: constants converted to 4×64-bit, loop bounds fixed
- KeyCodec: B constant fixed
- Glv: constants partially converted but regex left residual old values
- Point: types fixed, GX/GY constants converted
- Secp256k1: parameter types updated

Still needs: manual cleanup of Glv constants, ScalarN reduceWide internals,
test hex() helpers, test constant arrays, wNAF bit manipulation for 64-bit limbs.

+++ quartz/src/commonMain/kotlin/com/vitorpamplona/quartz/utils/secp256k1/Secp256k1.kt
@@ -227,7 +227,7 @@ object Secp256k1 {
      */
     private fun signSchnorrInternal(
         data: ByteArray,
-        d0: IntArray,
+        d0: LongArray,
         pBytes: ByteArray,
         pubKeyHasEvenY: Boolean,
         auxrand: ByteArray?,
