+++ quartz/src/commonTest/kotlin/com/vitorpamplona/quartz/utils/secp256k1/U256Test.kt
@@ -32,7 +32,7 @@ class U256Test {
             s.chunked(2).map { it.toInt(16).toByte() }.toByteArray(),
         )
 
-    private fun toHex(a: IntArray) = U256.toBytes(a).joinToString("") { "%02x".format(it) }
+    private fun toHex(a: LongArray) = U256.toBytes(a).joinToString("") { "%02x".format(it) }
 
     // ==================== isZero / cmp ====================
